[IMP] improve partner ledger views 1. movemos el cambio de accion de partner ledger de tax settlement a account accountant para que aun sin tax settlement ya cambie comportamiento 2. sacamos boton de partners que lleva a vista de partner ledger tipo tree y dejamos un solo boton que lleva al informe (modificado en punto 1, nativamente odoo a partir de 17 tmb lleva a una tree). Para ir a la tree se puede ir luego desde el informe 3. agregamos boton para ir a follow up desde nuestra "tree partner ledger"
