+++ account_accountant_ux/__manifest__.py
@@ -19,7 +19,7 @@
 ##############################################################################
 {
     'name': 'Accounting Accountant UX',
-    'version': "17.0.1.1.0",
+    'version': "17.0.1.2.0",
     'category': 'Accounting',
     'sequence': 14,
     'summary': '',
