+++ account_accountant_ux/models/__init__.py
@@ -9,3 +9,4 @@ from . import account_move_line
 from . import account_move
 from . import res_partner
 from . import account_journal_dashboard
+from . import account_partner_ledger
