+++ account_accountant_ux/views/account_move_line.xml
@@ -12,4 +12,15 @@
             </filter>
         </field>
     </record>
+
+    <record id="view_move_line_payment_tree" model="ir.ui.view">
+        <field name="name">account.move.line.inherit.view.tree</field>
+        <field name="model">account.move.line</field>
+        <field name="inherit_id" ref="account.view_move_line_payment_tree"/>
+        <field name="arch" type="xml">
+            <groupby name="partner_id">
+                <button type="object" name="open_action_followup" icon="fa-list" title="Follow Up"/>
+            </groupby>
+        </field>
+    </record>
 </odoo>
