+++ account_accountant_ux/views/res_partner_view.xml
@@ -6,23 +6,6 @@
         <field name="model">res.partner</field>
         <field name="inherit_id" ref="account_reports.partner_view_buttons"/>
         <field name="arch" type="xml">
-            <button id="partner_ledger_button" position="attributes">
-                <attribute name="icon">fa-file-text-o</attribute>
-                <attribute name="string">Partner Ledger Report</attribute>
-            </button>
-            <button id="partner_ledger_button" position="before">
-                <button
-                    type="action"
-                    class="oe_stat_button"
-                    id="partner_ledger_view_button"
-                    icon="fa-list"
-                    name="%(account.action_account_moves_ledger_partner)d"
-                    string="Partner Ledger"
-                    context="{'search_default_partner_id': id}"
-                    >
-                </button>
-            </button>
-
             <div name="button_box" position="inside">
                 <button class="oe_stat_button" type="object" groups="account.group_account_invoice"
                     name="action_open_reconcile" icon="fa-dollar">
