+++ account_tax_settlement/models/__init__.py
@@ -4,4 +4,3 @@ from . import account_move
 from . import account_journal_dashboard
 from . import account_report
 from . import account_generic_tax_report
-from . import res_partner
